requirements.txt
# Django Core
Django==4.2.7
asgiref==3.9.1
sqlparse==0.5.3
pytz==2025.2
typing_extensions==4.14.1

# REST API Support
djangorestframework==3.14.0
django-filter==23.4
drf-spectacular==0.26.5
uritemplate==4.2.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1

# CORS (for frontend integration like React/Next.js)
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Deployment
gunicorn==21.2.0

# Database Drivers
psycopg2-binary==2.9.9  # PostgreSQL
mysqlclient==2.2.0      # MySQL/MariaDB
PyMySQL==1.1.1

# Image Processing
Pillow==10.2.0

# CKEditor & Assets
django-ckeditor==6.7.0
django-js-asset==3.1.2

# Django Extensions
django-extensions==3.2.3

# Utilities
attrs==25.3.0
dj-database-url==2.1.0
inflection==0.5.1
requests>=2.0.0

# JSON & YAML
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
